found_mvp/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
httpx
